pyproject.toml
[project]
name = "realmate-challenge"
version = "0.1.0"
description = ""
authors = [{ name = "Pedro Henrique de Aquino", email = "aquinopedro36@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = ["django (>=5.1.6,<6.0.0)", "djangorestframework (>=3.16.1,<4.0.0)", "psycopg2-binary (>=2.9.11,<3.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-django = "^4.11.1"

